Add token to user model and move logic from authentication controller to service
