+++ KSPZL.Data/Models/User.cs
@@ -12,5 +12,6 @@ namespace KSZPL.Data.Models
         public byte[] PasswordHash { get; set; }
         public byte[] PasswordSalt { get; set; }
         public string Role { get; set; }
+        public string Token { get; set; }
     }
 }
