+++ KSZPL/Dtos/User/UserDto.cs
@@ -7,5 +7,6 @@
         public string Username { get; set; }
         public string Password { get; set; }
         public string Role { get; set; }
+        public string Token { get; set; }
     }
 }
